WebScraping/Upwork/Ahmad/active-jobs/Milestone27/requirements.txt
camelot_py==1.0.0
openpyxl==3.1.5
pandas==2.2.3
rich==13.9.4
